requirements.txt
flask
flask-cors
torch
torchvision
pillow
gunicorn
python-dotenv
requests
